+++ src/main/java/co/wanted/board/api/member/application/SignupService.java
@@ -1,6 +1,6 @@
 package co.wanted.board.api.member.application;
 
-import co.wanted.board.api.auth.encode.application.PasswordEncoder;
+import co.wanted.board.api.auth.application.encode.PasswordEncoder;
 import co.wanted.board.api.member.domain.Member;
 import co.wanted.board.api.member.domain.Password;
 import co.wanted.board.api.member.infrastructure.persistence.MemberRepository;
